+++ containers/daap-delete-table-for-data-product/tests/unit/conftest.py
@@ -1,6 +1,7 @@
 import json
 import os
 import sys
+import urllib
 from dataclasses import dataclass
 from os.path import dirname, join
 
@@ -93,7 +94,7 @@ def data_product_element(data_product_name, table_name):
 
 @pytest.fixture
 def region_name():
-    return "eu-west-1"
+    return "us-east-1"
 
 
 @pytest.fixture
@@ -139,14 +140,32 @@ def data_product_versions():
 
 @pytest.fixture
 def create_metadata(
-    s3_client, create_metadata_bucket, data_product_name, data_product_versions
+    s3_client,
+    create_metadata_bucket,
+    data_product_name,
+    data_product_versions,
+    table_name,
 ):
     for version in data_product_versions:
         s3_client.put_object(
-            Bucket=os.environ.get("METADATA_BUCKET", "metadata"),
+            Bucket=os.getenv("METADATA_BUCKET"),
             Key=f"{data_product_name}/{version}/metadata.json",
-            Body=json.dumps({"test": version}),
+            Body=json.dumps(
+                {
+                    "name": "test_product1",
+                    "description": "testing",
+                    "domain": "MoJ",
+                    "dataProductOwner": "matthew.laverty@justice.gov.uk",
+                    "dataProductOwnerDisplayName": "matt laverty",
+                    "email": "matthew.laverty@justice.gov.uk",
+                    "status": "draft",
+                    "dpiaRequired": False,
+                    "retentionPeriod": 3000,
+                    "schemas": [table_name],
+                }
+            ),
         )
+    load_v1_metadata_schema_to_mock_s3(s3_client)
 
 
 @pytest.fixture
@@ -201,3 +220,21 @@ def create_curated_data(
                 Key=f"curated/{data_product_name}/{version}/{table_name}/curated-file-{str(i)}.json",
                 Body=json.dumps({"content": f"{i}"}),
             )
+
+
+def load_v1_metadata_schema_to_mock_s3(s3_client):
+    with urllib.request.urlopen(
+        "https://raw.githubusercontent.com/ministryofjustice/modernisation-platform-environments/main/terraform/environments/data-platform/data-product-metadata-json-schema/v1.1.0/moj_data_product_metadata_spec.json"  # noqa E501
+    ) as url:
+        data = json.load(url)
+    json_data = json.dumps(data)
+    s3_client.put_object(
+        Body=json_data,
+        Bucket=os.environ["METADATA_BUCKET"],
+        Key="data_product_metadata_spec/v1.1.0/moj_data_product_metadata_spec.json",
+    )
+
+
+@pytest.fixture
+def regenerate_schema_data(create_schema):
+    pass