+++ containers/daap-athena-load/CHANGELOG.md
@@ -25,9 +25,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   falling back to inferred metadata if none is found
 - curated database name includes major version
 
+## [1.2.4] 2023-11-6
+
 ### Changed
 
-- Updated log level from error to info when table is first created
+- `athena_load_handler_test.py` updated s3 client reference to `s3_client`
 
 ## [1.2.3] 2023-10-18